requirements.txt
python-docx
openai-whisper
tqdm
catboost
sklearn
nlpaug
gensim